refactor(transaction-draft): split simple-mode lens into extension file (#272, #267)
Extract the computed accessors, editing methods, and mode-switching
helpers that operate on the simple-mode lens (`relevantLeg` /
`counterpartLeg`) into a sibling file
`Shared/Models/TransactionDraft+SimpleMode.swift`. The main
`TransactionDraft.swift` keeps the type definition, convenience
initialisers, validation, conversion, autofill, earmark-only
invariants, and custom-mode operations.

Drops `TransactionDraft.swift` from 586 → 379 lines, clearing its
`file_length` baseline entry. Two source-side adjustments needed to
satisfy SwiftLint without baseline edits:

- `counterpartLegIndex` was `private` and is now `internal` (no
  modifier). The strict_fileprivate rule disallows `fileprivate`,
  so internal is the smallest legal scope when the helper moves to
  a sibling file. Its surface widening is documented at the call
  site and it remains an implementation detail of the simple-mode
  helpers.
- `let a` / `let b` in `canSwitchToSimple` renamed to `primary` /
  `counterpart` to avoid re-keying baselined `identifier_name`
  entries under the new path.

No behavioural change.
